agentos/pyproject.toml
[project]
name = "agentos"
version = "0.2.0"
description = "12-factor aligned onboarding agent system"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "agno",
    "anthropic",
    "boto3",
    "fastapi[standard]",
    "python-dotenv",
    "langwatch",
    "openinference-instrumentation-agno",
    "sqlalchemy>=2.0.46",
    "pyjwt>=2.11.0",
]

[project.optional-dependencies]
postgres = [
    "psycopg[binary]>=3.1.0",
]
test = [
    "langwatch-scenario",
    "pytest",
    "pytest-asyncio",
    "pytest-asyncio-concurrent",
    "pandas",
    "httpx",
]

[tool.pytest.ini_options]
asyncio_mode = "auto"
markers = [
    "agent_test: marks tests as agent scenario tests",
]
